+++ src/main/java/com/epam/web/entitiy/User.java
@@ -1,14 +1,16 @@
 package com.epam.web.entitiy;
 
+import java.math.BigDecimal;
+
 public class User implements Identifiable {
 
     private final long id;
     private final String login;
     private final String password;
-    private final double balance;
+    private final BigDecimal balance;
     private final UserRole role;
 
-    public User(long id, String login, String password, double balance, UserRole role) {
+    public User(long id, String login, String password, BigDecimal balance, UserRole role) {
         this.id = id;
         this.login = login;
         this.password = password;
@@ -33,12 +35,8 @@ public class User implements Identifiable {
         return this.role.toString();
     }
 
-    public double getBalance() {
+    public BigDecimal getBalance() {
         return this.balance;
     }
-
-//    public DecimalFormat getBalance() {
-//        return this.balance;
-//    }
 
 }
